Move scopesSchema to oauth/validators.ts file
In the future other input validations could also be moved here
